requirements.txt
gdal
django==3.2.*
Pillow==9.1.*
django-cleanup
